requirements.txt
pandas
openpyxl
Pillow
fonttools
brotli
reportlab
